+++ core/modules/enqueuer.py
@@ -2,9 +2,11 @@ import argparse
 import json
 
 from datetime import datetime
-from database import get_engine_db
 from pika import BlockingConnection, ConnectionParameters
-from products import Product_db
+
+from core.db.database import get_engine_db
+from core.db.products import Product_db
+
 
 __author__ = 'lucas'
